Backend/requirements.txt
# Django Core
Django==5.2.4
asgiref==3.9.1

# Django Extensions and Tools
django-cors-headers==4.7.0
django-extensions==4.1

# Django Ninja API Framework
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7

# Database
psycopg2-binary==2.9.10

# Authentication & JWT
PyJWT==2.10.1

# Data Validation
pydantic==2.11.7
pydantic-core==2.33.2

# HTTP Requests
requests==2.32.4
urllib3==2.5.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10

# Image Processing
Pillow==11.3.0

# Configuration Management
python-decouple==3.8

# Cryptography
cryptography==45.0.5
cffi==1.17.1
pycparser==2.22

# Type Annotations
typing-extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# Dependency Injection
injector==0.22.0

# SQL Parsing
sqlparse==0.5.3

# Timezone Data
tzdata==2025.2

# Context Management
contextlib2==21.6.0
